# apps/backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service en bonne santé"
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/quizzes": {
      "get": {
        "operationId": "QuizzesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "post": {
        "operationId": "QuizzesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}": {
      "get": {
        "operationId": "QuizzesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "put": {
        "operationId": "QuizzesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "delete": {
        "operationId": "QuizzesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/status": {
      "patch": {
        "operationId": "QuizzesController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/api/v1/quizzes/{id}/questions": {
      "post": {
        "operationId": "QuestionsController_add",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "questions"
        ]
      }
    },
    "/api/v1/quizzes/{id}/questions/reorder": {
      "patch": {
        "operationId": "QuestionsController_reorder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderQuestionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "questions"
        ]
      }
    },
    "/api/v1/questions/{qid}": {
      "put": {
        "operationId": "QuestionsController_update",
        "parameters": [
          {
            "name": "qid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "questions"
        ]
      },
      "delete": {
        "operationId": "QuestionsController_remove",
        "parameters": [
          {
            "name": "qid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "questions"
        ]
      }
    }
  },
  "info": {
    "title": "Roux-Quizz API",
    "description": "API REST du builder de quiz et des restitutions",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "host",
          "player",
          "admin"
        ],
        "description": "Rôle."
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant interne (ULID)."
          },
          "displayName": {
            "type": "string",
            "description": "Nom affiché."
          },
          "email": {
            "type": "object",
            "description": "Courriel, si connu.",
            "nullable": true
          },
          "role": {
            "description": "Rôle.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "role"
        ]
      },
      "QuizDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "unlisted"
            ]
          },
          "language": {
            "type": "string"
          },
          "questionCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "ownerId",
          "title",
          "description",
          "coverMediaId",
          "status",
          "visibility",
          "language",
          "questionCount",
          "createdAt",
          "updatedAt",
          "archivedAt"
        ]
      },
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "language": {
            "default": "fr",
            "type": "string",
            "minLength": 2,
            "maxLength": 5
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26
          }
        },
        "required": [
          "title"
        ]
      },
      "QuizDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverMediaId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "unlisted"
            ]
          },
          "language": {
            "type": "string"
          },
          "questionCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "archivedAt": {
            "type": "string",
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "quizId": {
                  "type": "string"
                },
                "orderIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "single_choice",
                    "multiple_choice",
                    "true_false",
                    "text_input",
                    "numeric",
                    "ordering",
                    "poll"
                  ]
                },
                "prompt": {
                  "type": "string"
                },
                "mediaId": {
                  "type": "string",
                  "nullable": true
                },
                "timeLimitS": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "pointsMode": {
                  "type": "string",
                  "enum": [
                    "standard",
                    "double",
                    "none"
                  ]
                },
                "numericValue": {
                  "type": "string",
                  "nullable": true
                },
                "numericTolerance": {
                  "type": "string",
                  "nullable": true
                },
                "options": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "orderIndex": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "text": {
                        "type": "string",
                        "nullable": true
                      },
                      "mediaId": {
                        "type": "string",
                        "nullable": true
                      },
                      "color": {
                        "type": "string",
                        "enum": [
                          "red",
                          "blue",
                          "yellow",
                          "green"
                        ]
                      },
                      "shape": {
                        "type": "string",
                        "enum": [
                          "triangle",
                          "diamond",
                          "circle",
                          "square"
                        ]
                      },
                      "isCorrect": {
                        "type": "boolean"
                      },
                      "correctOrderIndex": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991,
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "orderIndex",
                      "text",
                      "mediaId",
                      "color",
                      "shape",
                      "isCorrect",
                      "correctOrderIndex"
                    ]
                  }
                },
                "acceptedAnswers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "text": {
                        "type": "string"
                      },
                      "normalized": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "text",
                      "normalized"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "quizId",
                "orderIndex",
                "type",
                "prompt",
                "mediaId",
                "timeLimitS",
                "pointsMode",
                "numericValue",
                "numericTolerance",
                "options",
                "acceptedAnswers"
              ]
            }
          }
        },
        "required": [
          "id",
          "ownerId",
          "title",
          "description",
          "coverMediaId",
          "status",
          "visibility",
          "language",
          "questionCount",
          "createdAt",
          "updatedAt",
          "archivedAt",
          "questions"
        ]
      },
      "UpdateQuizDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "language": {
            "type": "string",
            "minLength": 2,
            "maxLength": 5
          },
          "coverMediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26,
            "nullable": true
          }
        }
      },
      "TransitionQuizDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "single_choice",
              "multiple_choice",
              "true_false",
              "text_input",
              "numeric",
              "ordering",
              "poll"
            ]
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "mediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26
          },
          "timeLimitS": {
            "default": 20,
            "type": "integer",
            "minimum": 5,
            "maximum": 120
          },
          "pointsMode": {
            "default": "standard",
            "type": "string",
            "enum": [
              "standard",
              "double",
              "none"
            ]
          },
          "numericValue": {
            "type": "number"
          },
          "numericTolerance": {
            "type": "number",
            "minimum": 0
          },
          "options": {
            "default": [],
            "maxItems": 6,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "maxLength": 500
                },
                "mediaId": {
                  "type": "string",
                  "minLength": 26,
                  "maxLength": 26
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "red",
                    "blue",
                    "yellow",
                    "green"
                  ]
                },
                "shape": {
                  "type": "string",
                  "enum": [
                    "triangle",
                    "diamond",
                    "circle",
                    "square"
                  ]
                },
                "isCorrect": {
                  "default": false,
                  "type": "boolean"
                },
                "correctOrderIndex": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "color",
                "shape"
              ]
            }
          },
          "acceptedAnswers": {
            "default": [],
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                }
              },
              "required": [
                "text"
              ]
            }
          }
        },
        "required": [
          "type",
          "prompt"
        ]
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "quizId": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "type": {
            "type": "string",
            "enum": [
              "single_choice",
              "multiple_choice",
              "true_false",
              "text_input",
              "numeric",
              "ordering",
              "poll"
            ]
          },
          "prompt": {
            "type": "string"
          },
          "mediaId": {
            "type": "string",
            "nullable": true
          },
          "timeLimitS": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pointsMode": {
            "type": "string",
            "enum": [
              "standard",
              "double",
              "none"
            ]
          },
          "numericValue": {
            "type": "string",
            "nullable": true
          },
          "numericTolerance": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "orderIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "text": {
                  "type": "string",
                  "nullable": true
                },
                "mediaId": {
                  "type": "string",
                  "nullable": true
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "red",
                    "blue",
                    "yellow",
                    "green"
                  ]
                },
                "shape": {
                  "type": "string",
                  "enum": [
                    "triangle",
                    "diamond",
                    "circle",
                    "square"
                  ]
                },
                "isCorrect": {
                  "type": "boolean"
                },
                "correctOrderIndex": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "nullable": true
                }
              },
              "required": [
                "id",
                "orderIndex",
                "text",
                "mediaId",
                "color",
                "shape",
                "isCorrect",
                "correctOrderIndex"
              ]
            }
          },
          "acceptedAnswers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                },
                "normalized": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "text",
                "normalized"
              ]
            }
          }
        },
        "required": [
          "id",
          "quizId",
          "orderIndex",
          "type",
          "prompt",
          "mediaId",
          "timeLimitS",
          "pointsMode",
          "numericValue",
          "numericTolerance",
          "options",
          "acceptedAnswers"
        ]
      },
      "ReorderQuestionsDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "questionId": {
                  "type": "string",
                  "minLength": 26,
                  "maxLength": 26
                },
                "orderIndex": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "questionId",
                "orderIndex"
              ]
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "single_choice",
              "multiple_choice",
              "true_false",
              "text_input",
              "numeric",
              "ordering",
              "poll"
            ]
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "mediaId": {
            "type": "string",
            "minLength": 26,
            "maxLength": 26
          },
          "timeLimitS": {
            "default": 20,
            "type": "integer",
            "minimum": 5,
            "maximum": 120
          },
          "pointsMode": {
            "default": "standard",
            "type": "string",
            "enum": [
              "standard",
              "double",
              "none"
            ]
          },
          "numericValue": {
            "type": "number"
          },
          "numericTolerance": {
            "type": "number",
            "minimum": 0
          },
          "options": {
            "default": [],
            "maxItems": 6,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "maxLength": 500
                },
                "mediaId": {
                  "type": "string",
                  "minLength": 26,
                  "maxLength": 26
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "red",
                    "blue",
                    "yellow",
                    "green"
                  ]
                },
                "shape": {
                  "type": "string",
                  "enum": [
                    "triangle",
                    "diamond",
                    "circle",
                    "square"
                  ]
                },
                "isCorrect": {
                  "default": false,
                  "type": "boolean"
                },
                "correctOrderIndex": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "color",
                "shape"
              ]
            }
          },
          "acceptedAnswers": {
            "default": [],
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                }
              },
              "required": [
                "text"
              ]
            }
          }
        },
        "required": [
          "type",
          "prompt"
        ]
      }
    }
  }
}